seperate brick controller.py & brick controller of ovs
